refactor(launcher): extract pool handlers into reducer/pool.rs (task #182 PR-B)
Extracted 5 pool-related handler functions from reducer/mod.rs into a
new reducer/pool.rs sub-module. Pure mechanical move; no logic
changes.

Extracted:
- handle_report_host_pool_count
- handle_report_pool_window_added
- handle_report_pool_window_removed
- handle_report_pool_window_promoted
- handle_report_pool_drain_decision

All marked pub(super) — only mod.rs's update() dispatcher calls them.
Dispatcher arms updated to `pool::handle_*` calls.

Sizes:
- agentmux-launcher/src/reducer/mod.rs: 1114 → 1007 lines (-107)
- agentmux-launcher/src/reducer/pool.rs: NEW 121 lines

Pattern established for follow-up extractions (window_lifecycle, saga,
connection, query). Each domain extraction will be a separate PR
following this same template.

154 launcher tests pass.
